+++ shared/pwa/serwist.ts
@@ -1,0 +1,4 @@
+'use client';
+
+//NOTE: This file is only used to export the SerwistProvider for use in the layout.tsx file, becouse client components cant be in node_modules.
+export { SerwistProvider } from '@serwist/turbopack/react';