+++ shared/pwa/sw.ts
@@ -1,0 +1,35 @@
+/// <reference no-default-lib="true" />
+/// <reference lib="esnext" />
+/// <reference lib="webworker" />
+import { defaultCache } from '@serwist/turbopack/worker';
+import type { PrecacheEntry, SerwistGlobalConfig } from 'serwist';
+import { NetworkFirst, Serwist, disableDevLogs } from 'serwist';
+
+disableDevLogs();
+
+declare global {
+	interface WorkerGlobalScope extends SerwistGlobalConfig {
+		__SW_MANIFEST: (PrecacheEntry | string)[] | undefined;
+	}
+}
+
+declare const self: ServiceWorkerGlobalScope;
+
+const serwist = new Serwist({
+	precacheEntries: self.__SW_MANIFEST,
+	skipWaiting: true,
+	clientsClaim: true,
+	navigationPreload: true,
+	runtimeCaching: [
+		{
+			matcher: /^\/api\/excersises/,
+			handler: new NetworkFirst({
+				cacheName: 'exercises-api',
+				networkTimeoutSeconds: 5,
+			}),
+		},
+		...defaultCache,
+	],
+});
+
+serwist.addEventListeners();